requirements.txt
aiohttp==3.9.5
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
